+++ include/WiFiManager.h
@@ -2,23 +2,20 @@
 #define WIFI_MANAGER_H
 
 #include <WiFi.h>
-#include <WebServer.h>
 #include <Preferences.h>
 
 class WiFiManager {
 private:
     Preferences preferences;
-    WebServer* server;
     String ssid;
     String password;
-    bool apMode;
     unsigned long connectTimeout;
     
-    void startAP();
-    void setupWebServer();
-    void handleRoot();
-    void handleSave();
-    void handleStatus();
+    // Predefined hotspot credentials for initial setup
+    const char* DEFAULT_SSID = "EspSetup";
+    const char* DEFAULT_PASSWORD = "HeLooWod";
+    
+    bool connectToWiFi(const String& ssid, const String& password);
 
 public:
     WiFiManager();
@@ -28,7 +25,6 @@ public:
     void loop();
     bool isConnected();
     bool isAPMode();
-    void startConfigPortal();
     bool loadCredentials();
     void saveCredentials(const String& ssid, const String& password);
     void clearCredentials();